requirements.txt
ollama==0.4.7
pandas==2.2.3
pydantic==2.11.1
PyYAML==6.0.2
ragflow_sdk==0.17.2
openpyxl==3.1.5
